requirements.txt
streamlit
ultralytics
opencv-python-headless
pillow
torch
matplotlib
numpy
grad-cam
